Add Pagination and Live Search Features

+++ delete.php
@@ -1,8 +1,8 @@
 
+
+<!-- Delete Post -->
 <?php
-
 session_start();
-
 if (!isset($_SESSION['email']))
     {
     echo "<script>alert('Plesase Login ');</script>";
@@ -13,6 +13,15 @@ if (!isset($_SESSION['email']))
 include 'src/db.php';
 include 'src/config.php';
 $id = $_GET['id'];
-$conn->query("DELETE FROM posts WHERE ID=$id");
+$post = $conn->query("SELECT * FROM posts WHERE ID=$id")->fetch_assoc();
+if ($post) {
+    //Remove image file from uploads folder ---------------------------
+    $image = $post['post_image'];
+    if ($image != "default.png") {
+        unlink("uploads/posts/$image");
+    }
+    //Delete post from database --------------------------------------
+    $conn->query("DELETE FROM posts WHERE ID=$id");
+}
 header("Location:admin");
 ?>
